+++ src/feature/auth/ui/SignIn/index.ts
@@ -1,0 +1,1 @@
+export * from './SignIn';
